feat(web): rebuild setup page on the shared Petrol Mono auth shell
Extract the login page's design into shared components under
web/src/components/auth/ (AuthShell two-column layout, ProductShowcase
demo loop, AuthField/AuthSubmitButton/AuthErrorAlert form primitives)
and rebuild /setup on top of them so it matches /auth. Both pages now
use getApiErrorMessage instead of hand-rolled axios error unwraps.
