+++ web/src/app/setup/layout.tsx
@@ -4,8 +4,6 @@ export default function SetupLayout({
 	children: React.ReactNode;
 }) {
 	return (
-		<div className="fixed inset-0 flex items-center justify-center bg-background p-4 overflow-hidden">
-			{children}
-		</div>
+		<div className="fixed inset-0 overflow-auto bg-marketing">{children}</div>
 	);
 }
